+++ packages/ui/package.json
@@ -1,99 +1,101 @@
 {
-	"name": "@nuclear/ui",
-	"version": "0.6.43",
-	"description": "Nuclear - UI components",
-	"author": "nukeop <12746779+nukeop@users.noreply.github.com>",
-	"homepage": "https://github.com/nukeop/nuclear/tree/master/packages/ui#readme",
-	"license": "AGPL-3.0",
-	"files": [
-		"lib"
-	],
-	"main": "dist/index.js",
-	"typings": "dist/index.d.ts",
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/nukeop/nuclear.git"
-	},
-	"bugs": {
-		"url": "https://github.com/nukeop/nuclear/issues"
-	},
-	"scripts": {
-		"storybook": "storybook dev -p 9001 -c .storybook",
-		"lint": "eslint \"{lib,stories}/**/*.{ts,tsx,js,jsx}\"",
-		"lint:fix": "eslint \"{lib,stories}/**/*.{ts,tsx,js,jsx}\" --fix",
-		"test": "jest",
-		"coverage": "jest --coverage"
-	},
-	"peerDependencies": {
-		"react": "16.14.0",
-		"react-dom": "16.14.0"
-	},
-	"devDependencies": {
-		"@babel/core": "7.26.7",
-		"@babel/plugin-proposal-object-rest-spread": "7.20.7",
-		"@babel/polyfill": "7.12.1",
-		"@babel/preset-env": "7.26.7",
-		"@babel/preset-react": "7.26.3",
-		"@babel/preset-typescript": "7.26.0",
-		"@babel/register": "7.25.9",
-		"@storybook/addon-actions": "8.6.6",
-		"@storybook/addon-info": "5.3.21",
-		"@storybook/addon-knobs": "8.0.1",
-		"@storybook/addon-links": "8.6.6",
-		"@storybook/addon-webpack5-compiler-babel": "^3.0.5",
-		"@storybook/cli": "8.6.6",
-		"@storybook/preset-typescript": "3.0.0",
-		"@storybook/react": "8.6.6",
-		"@storybook/react-webpack5": "^8.6.6",
-		"@storybook/test": "^8.6.6",
-		"@testing-library/react": "^11.2.7",
-		"@types/jest": "^27.4.1",
-		"@types/lodash": "^4.14.158",
-		"@types/numeral": "0.0.28",
-		"@types/react-beautiful-dnd": "^13.0.0",
-		"@types/react-measure": "^2.0.6",
-		"@types/react-table": "^7.7.16",
-		"babel-eslint": "10.1.0",
-		"babel-loader": "9.2.1",
-		"copyfiles": "^2.2.0",
-		"css-loader": "7.1.2",
-		"eslint": "^7.18.0",
-		"fork-ts-checker-webpack-plugin": "^6.2.5",
-		"identity-obj-proxy": "^3.0.0",
-		"ignore-styles": "^5.0.1",
-		"jest": "^27.5.1",
-		"react-hot-loader": "4.13.1",
-		"sass": "1.83.4",
-		"sass-loader": "16.0.4",
-		"style-loader": "4.0.0",
-		"ts-jest": "^27.1.3",
-		"ts-loader": "9.5.2",
-		"ts-node": "^10.7.0",
-		"typescript": "^4.2.4"
-	},
-	"dependencies": {
-		"@nuclear/core": "^0.6.43",
-		"@types/react-window": "^1.8.5",
-		"butterchurn": "^2.6.7",
-		"butterchurn-presets": "^2.4.7",
-		"classnames": "^2.2.6",
-		"formik": "^2.2.9",
-		"history": "^4.10.1",
-		"lodash": "^4.17.15",
-		"numeral": "^2.0.6",
-		"prop-types": "^15.6.2",
-		"react-beautiful-dnd": "^13.1.0",
-		"react-cool-onclickoutside": "^1.7.0",
-		"react-full-screen": "^1.0.1",
-		"react-image": "^2.2.2",
-		"react-measure": "^2.5.2",
-		"react-resizable": "^3.0.4",
-		"react-table": "^7.8.0",
-		"react-virtualized-auto-sizer": "^1.0.15",
-		"react-window": "^1.8.8",
-		"recompose": "^0.30.0",
-		"semantic-ui-react": "2.1.5",
-		"styled-components": "^4.4.1",
-		"yup": "^0.32.9"
-	}
+  "name": "@nuclear/ui",
+  "version": "0.6.43",
+  "description": "Nuclear - UI components",
+  "author": "nukeop <12746779+nukeop@users.noreply.github.com>",
+  "homepage": "https://github.com/nukeop/nuclear/tree/master/packages/ui#readme",
+  "license": "AGPL-3.0",
+  "files": [
+    "lib"
+  ],
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nukeop/nuclear.git"
+  },
+  "bugs": {
+    "url": "https://github.com/nukeop/nuclear/issues"
+  },
+  "scripts": {
+    "storybook": "storybook dev -p 9001 -c .storybook",
+    "lint": "eslint \"{lib,stories}/**/*.{ts,tsx,js,jsx}\"",
+    "lint:fix": "eslint \"{lib,stories}/**/*.{ts,tsx,js,jsx}\" --fix",
+    "test": "jest",
+    "coverage": "jest --coverage"
+  },
+  "peerDependencies": {
+    "react": "16.14.0",
+    "react-dom": "16.14.0"
+  },
+  "devDependencies": {
+    "@babel/core": "7.26.7",
+    "@babel/plugin-proposal-object-rest-spread": "7.20.7",
+    "@babel/polyfill": "7.12.1",
+    "@babel/preset-env": "7.26.7",
+    "@babel/preset-react": "7.26.3",
+    "@babel/preset-typescript": "7.26.0",
+    "@babel/register": "7.25.9",
+    "@storybook/addon-actions": "^8.6.7",
+    "@storybook/addon-info": "5.3.21",
+    "@storybook/addon-knobs": "8.0.1",
+    "@storybook/addon-links": "^8.6.7",
+    "@storybook/addon-styling-webpack": "^1.0.1",
+    "@storybook/addon-webpack5-compiler-babel": "^3.0.5",
+    "@storybook/cli": "^8.6.7",
+    "@storybook/preset-scss": "^1.0.3",
+    "@storybook/preset-typescript": "3.0.0",
+    "@storybook/react": "^8.6.7",
+    "@storybook/react-webpack5": "^8.6.7",
+    "@storybook/test": "^8.6.7",
+    "@testing-library/react": "^11.2.7",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.158",
+    "@types/numeral": "0.0.28",
+    "@types/react-beautiful-dnd": "^13.0.0",
+    "@types/react-measure": "^2.0.6",
+    "@types/react-table": "^7.7.16",
+    "babel-eslint": "10.1.0",
+    "babel-loader": "9.2.1",
+    "copyfiles": "^2.2.0",
+    "css-loader": "7.1.2",
+    "eslint": "^7.18.0",
+    "fork-ts-checker-webpack-plugin": "^6.2.5",
+    "identity-obj-proxy": "^3.0.0",
+    "ignore-styles": "^5.0.1",
+    "jest": "^27.5.1",
+    "react-hot-loader": "4.13.1",
+    "sass": "1.83.4",
+    "sass-loader": "16.0.4",
+    "style-loader": "4.0.0",
+    "ts-jest": "^27.1.3",
+    "ts-loader": "9.5.2",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.2.4"
+  },
+  "dependencies": {
+    "@nuclear/core": "^0.6.43",
+    "@types/react-window": "^1.8.5",
+    "butterchurn": "^2.6.7",
+    "butterchurn-presets": "^2.4.7",
+    "classnames": "^2.2.6",
+    "formik": "^2.2.9",
+    "history": "^4.10.1",
+    "lodash": "^4.17.15",
+    "numeral": "^2.0.6",
+    "prop-types": "^15.6.2",
+    "react-beautiful-dnd": "^13.1.0",
+    "react-cool-onclickoutside": "^1.7.0",
+    "react-full-screen": "^1.0.1",
+    "react-image": "^2.2.2",
+    "react-measure": "^2.5.2",
+    "react-resizable": "^3.0.4",
+    "react-table": "^7.8.0",
+    "react-virtualized-auto-sizer": "^1.0.15",
+    "react-window": "^1.8.8",
+    "recompose": "^0.30.0",
+    "semantic-ui-react": "2.1.5",
+    "styled-components": "^4.4.1",
+    "yup": "^0.32.9"
+  }
 }
